GUI app, CLI and reader firmware  bug fixes

+++ GameBoyReader/GameBoyReader.Core/Models/CartridgeInformation.cs
@@ -1,9 +1,4 @@
 ﻿using GameBoyReader.Core.Enums;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace GameBoyReader.Core.Models
 {
@@ -12,6 +7,6 @@ namespace GameBoyReader.Core.Models
         public string Name { get; set; } = "";
         public CartridgeType Type { get; set; }
         public Byte ROMSize { get; set; }
-        public Byte RAMSize { get; set; }
+        public int RAMSize { get; set; }
     }
 }
